Updated: minor improvs

+++ include/ink/ink_base.hpp
@@ -128,14 +128,14 @@
 #include <cstring>
 
 using i8 = signed char;
-using i16 = signed short int;
+using i16 = signed short;
 using i32 = signed int;
-using i64 = signed long int;
+using i64 = signed long long;
 
 using u8 = unsigned char;
-using u16 = unsigned short int;
+using u16 = unsigned short;
 using u32 = unsigned int;
-using u64 = unsigned long int;
+using u64 = unsigned long long;
 
 using usize = size_t;
 using isize = std::ptrdiff_t;
